+++ src/generators/scanner_config.rs
@@ -8,7 +8,7 @@ use std::fmt::{Debug, Display, Error, Formatter};
 // ---------------------------------------------------
 ///
 /// Configuration information for a scanner.
-/// Contains features like to optionally switch automatic handling off newlines off.
+/// Contains features like to optionally switch automatic handling off and newlines off.
 ///
 #[derive(Debug, Clone, Eq, PartialEq, Ord, PartialOrd)]
 pub struct ScannerConfig {